+++ spec/command_line_spec.rb
@@ -2,4 +2,5 @@
 
 RSpec.describe 'CommandLine' do
   include_examples 'command_line', CommandLine.method(:command_line)
+  include_examples 'command_line!', CommandLine.method(:command_line!)
 end