+++ spec/fixtures/exit_status_failure.rb
@@ -1,1 +1,2 @@
+$stderr.print 'Error!'
 exit false