+++ spec/include_command_line_spec.rb
@@ -4,4 +4,5 @@ require 'command_line/global'
 
 RSpec.describe 'globally includes command_line' do
   include_examples 'command_line', method(:command_line)
+  include_examples 'command_line!', method(:command_line!)
 end