+++ spec/require_comand_line_include_spec.rb
@@ -4,4 +4,5 @@ RSpec.describe 'manually includes CommandLine to make command_line available' do
   include CommandLine
 
   include_examples 'command_line', method(:command_line)
+  include_examples 'command_line!', method(:command_line!)
 end